Initial FemWorkbench refraction towards CaeWorkbench see /Fem/Readme.md for details, C++ untouched
