+++ src/Mod/Fem/Cae.dox
@@ -1,3 +1,3 @@
-/** \defgroup FEM Fem
+/** \defgroup TEMPLATE Cfd
  *  \ingroup WORKBENCHES */
