+++ src/Mod/Fem/Init.py
@@ -1,29 +1,4 @@
-# FreeCAD init script of the Fem module  
-# (c) 2001 Juergen Riegel
 
-#***************************************************************************
-#*   (c) Juergen Riegel (juergen.riegel@web.de) 2002                       *   
-#*                                                                         *
-#*   This file is part of the FreeCAD CAx development system.              *
-#*                                                                         *
-#*   This program is free software; you can redistribute it and/or modify  *
-#*   it under the terms of the GNU Lesser General Public License (LGPL)    *
-#*   as published by the Free Software Foundation; either version 2 of     *
-#*   the License, or (at your option) any later version.                   *
-#*   for detail see the LICENCE text file.                                 *
-#*                                                                         *
-#*   FreeCAD is distributed in the hope that it will be useful,            *
-#*   but WITHOUT ANY WARRANTY; without even the implied warranty of        *
-#*   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the         *
-#*   GNU Lesser General Public License for more details.                   *
-#*                                                                         *
-#*   You should have received a copy of the GNU Library General Public     *
-#*   License along with FreeCAD; if not, write to the Free Software        *
-#*   Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  *
-#*   USA                                                                   *
-#*                                                                         *
-#*   Juergen Riegel 2002                                                   *
-#***************************************************************************/
 
 
 FreeCAD.addExportType("TetGen file (*.poly)","convert2TetGen") 
@@ -31,3 +6,28 @@ FreeCAD.addImportType("FEM formats (*.unv *.med *.dat *.bdf)","Fem")
 FreeCAD.addExportType("FEM formats (*.unv *.med *.dat *.inp)","Fem")
 FreeCAD.addImportType("CalculiX result (*.frd)","ccxFrdReader")
 FreeCAD.addImportType("Abaqus file (*.inp)","FemGui")
+
+#ls *.py
+#CaeAnalysis.py
+
+#CaeSolver.py
+#femCommands.py
+
+#ccxFemSolver.py
+#ccxFrdReader.py
+#ccxInpWriter.py
+#convert2TetGen.py
+#foamCfdSolver.py
+#foamResultReader.py
+#foamCaseWriter.py
+
+#cfdCommands_notNeeded.py
+#cfdResultControlTaskPanel.py
+
+
+#MechanicalMaterial.py
+
+#TestCfd.py
+#FemExample.py
+#TestFem.py
+
