+++ src/Mod/Fem/foamCaseWriter.py
@@ -1,0 +1,9 @@
+import FreeCAD
+import os
+import sys
+import time
+
+
+def is_solid_mesh(fem_mesh):
+    if fem_mesh.VolumeCount > 0:  # solid mesh
+        return True
