+++ src/Mod/Fem/foamResultReader.py
@@ -1,0 +1,54 @@
+#***************************************************************************
+#*                                                                         *
+#*   Copyright (c) 2013 - Joachim Zettler                                  *
+#*   Copyright (c) 2013 - Juergen Riegel <FreeCAD@juergen-riegel.net>      *
+#*                                                                         *
+#*   This program is free software; you can redistribute it and/or modify  *
+#*   it under the terms of the GNU Lesser General Public License (LGPL)    *
+#*   as published by the Free Software Foundation; either version 2 of     *
+#*   the License, or (at your option) any later version.                   *
+#*   for detail see the LICENCE text file.                                 *
+#*                                                                         *
+#*   This program is distributed in the hope that it will be useful,       *
+#*   but WITHOUT ANY WARRANTY; without even the implied warranty of        *
+#*   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the         *
+#*   GNU Library General Public License for more details.                  *
+#*                                                                         *
+#*   You should have received a copy of the GNU Library General Public     *
+#*   License along with this program; if not, write to the Free Software   *
+#*   Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  *
+#*   USA                                                                   *
+#*                                                                         *
+#***************************************************************************
+
+
+import FreeCAD
+import os
+from math import pow, sqrt
+
+__title__ = "FreeCAD Calculix library"
+__author__ = "Juergen Riegel "
+__url__ = "http://www.freecadweb.org"
+
+if open.__module__ == '__builtin__':
+    pyopen = open  # because we'll redefine open below
+
+def convert_foam():
+    pass
+
+
+def insert(filename, docname):
+    "called when freecad wants to import a file"
+    try:
+        doc = FreeCAD.getDocument(docname)
+    except NameError:
+        doc = FreeCAD.newDocument(docname)
+    FreeCAD.ActiveDocument = doc
+
+    importFrd(filename)
+
+
+def open(filename):
+    "called when freecad opens a file"
+    docname = os.path.splitext(os.path.basename(filename))[0]
+    insert(filename, docname)
